GH-917 Finish with JSONRPC_2_0
